+++ ChangeLog.md
@@ -18,8 +18,13 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-3.2.0
------
+3.1.15
+------
+
+3.1.14 - 06/20/2022
+-------------------
+- emcc now accepts `-print-file-name` and reports the correct library paths in
+  `-print-search-dirs`.
 - `tools/file_packager` no longer generates (or requires) any "pre-js" code when
   running in `--embed-file` mode.  Instead the embedded files are loaded at
   static constructor time.